requirements.txt
fastapi
uvicorn
python-dotenv
httpx
jinja2
python-multipart
sqlalchemy
